Ajout de la vérification du statut de l'utilisateur

+++ src/Controller/ArticleController.php
@@ -24,6 +24,11 @@ class ArticleController
 
     public function new()
     {
+        if (!isset($_SESSION['user'])) {
+            header('Location: /');
+            die;
+        }
+
         /**
          * Tableau d'arguments qui va nous permettre de récupérer les données souhaitées dans filter_input_array
          * Les données qu'on souhaite récupérer sont : "title" et "content"
@@ -76,6 +81,11 @@ class ArticleController
 
     public function edit(int $id)
     {
+        if (!isset($_SESSION['user'])) {
+            header('Location: /');
+            die;
+        }
+
         $articleDao = new ArticleDao();
         $article = $articleDao->getById($id);
 
@@ -121,6 +131,11 @@ class ArticleController
 
     public function delete(int $id)
     {
+        if (!isset($_SESSION['user'])) {
+            header('Location: /');
+            die;
+        }
+
         $articleDao = new ArticleDao();
         $articleDao->delete($id);
         header('Location: /');
